+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add support to CSI Spec `v0.3.0`. This includes many new changes, make sure 
   to read the Github PR for more information
   [[GH-72]](https://github.com/digitalocean/csi-digitalocean/pull/72)
+* Check volume limits before provisioning calls
+  [[GH-73]](https://github.com/digitalocean/csi-digitalocean/pull/73)
 
 ## v0.1.5 - 2018.08.27